refactor: slice 분리하여 관리
