+++ redexCounter/src/components/Header.js
@@ -1,5 +1,5 @@
 import { useDispatch, useSelector } from 'react-redux';
-import { authActions } from '../store';
+import { authActions } from '../store/authSlice';
 import classes from './Header.module.css';
 
 const Header = () => {
